+++ Magnesium.Vulkan/Structs/VkImageCreateInfo.cs
@@ -20,7 +20,7 @@ namespace Magnesium.Vulkan
 		public VkImageUsageFlags usage { get; set; }
 		public VkSharingMode sharingMode { get; set; }
 		public UInt32 queueFamilyIndexCount { get; set; }
-		public UInt32 pQueueFamilyIndices { get; set; }
+		public IntPtr pQueueFamilyIndices { get; set; }
 		public VkImageLayout initialLayout { get; set; }
 	}
 }
